# schemahcl: reduce eval mem usage