```
benchmark              old ns/op     new ns/op     delta
BenchmarkSchema-16     569319394     238469401     -58.11%

benchmark              old allocs     new allocs     delta
BenchmarkSchema-16     1327927        1174085        -11.59%

benchmark              old bytes     new bytes     delta
BenchmarkSchema-16     293956792     186829161     -36.44%
```